Bugfix in linking maint. for panel window move (again)
This time the issue was when a target (not source) panel was the one that's moving.

When the moved panel closed in its old window, we purged its uuid as a target from
all linking maps, thus unlinking it, which we did not want to do.

So now, when any panel is closing, the `GlobalLinkingMap` handler for the event checks
to see if the uuid still exists in some other window. Iff it does, then we decline to purge the
uuid as a target.

+++ client/src/content_types/linking.js
@@ -295,7 +295,13 @@ export class GlobalLinkingMap {
         this.localComponent = new LinkingMapComponent(this.hub);
         this.hub.windowManager.addHandler(this.name, this.localComponent);
         this.hub.contentManager.on('localPaneClose', async event => {
-            await this.purgeTarget(event.uuid);
+            // Purge the target only if its uuid is absent from any other window. It will be present
+            // in another window if we are in the process of moving it to that window.
+            const existsElsewhere = await this.hub.contentManager.uuidExistsInAnyWindow(
+                event.uuid, {excludeSelf: true});
+            if (!existsElsewhere) {
+                await this.purgeTarget(event.uuid);
+            }
             this.localComponent.deleteForAllX({u: event.uuid});
         });
         this.hub.contentManager.on('paneMovedToAnotherWindow', async event => {
